requirements.txt
kafka-python==2.2.15
qdrant-client==1.16.2
requests==2.32.5
pydantic==2.12.5
numpy==2.3.3
tqdm==4.67.1
geopy==2.4.1
sentence-transformers==5.2.0
transformers==4.57.3
torch==2.8.0
pandas==2.3.3
PyYAML==6.0.3
wget==3.2
scikit-learn==1.7.2
scipy==1.16.2
openai==2.14.0

